Make a component file to export and insert in another file.
I learnt that you can make components in other files and import it into different files just by calling a class. Given it uses ES6 syntax, I also have learnt that the current format of the function in Person.js benefits this which is another discovery to seek. However it basically works.