Learn about props
Props can be named anything but currently are used as properties. This is because you can put arguments into HTML elements and then call props, to group those arguments to call. In a html element, you call the props via the key and then the result is the value. For instance name="Hello", name is the key, Hello is value.