Learnt about state
Instead of permenantly fixing the values, I can place it all in one place called state. Currently, state is an array of objects consisting of the name and age. The outcome is the same but the way we use this and state to call the objects in the array is formatted differently. If state is changed, React will re-render the Dom/update the DOm