Make a button switch the name for setState
setState allows us to not change the actual state but merge it with the state instead. For React in roder to chagne the state you need to use this.setState. In our code, when you click the button, the state for one of the name changes given that method is linked to an onclick method in the button. Whenever a state or prop is changed, React will analyze the what is thecurrent state, new state and render the outcome/merged state (I.E the new state)