How to pass in arguments with the bind method
From what i know, bind is required to call methods in order for methods to be called and used. You can also pass in arguements to bind methods if you wanted to change the name. The first line where its a one line function immediately gives a return but its not the best way to pass in an argument. Also, we have given switchNameHandler an argument so whatever name that gets passed in will be passed through the new State. Cool.