Learn two way binding.
In order to pass in methods and values, we are making input variables, our own input variables to pass it in to a different class. What's interesting is that how we are showing the current value of the name and the new name is two way binding, by binding more than one input element and value to the input type in the Person.js class. This is two way binding and you can change the name dynamically for the second name.