Dynamically changing state with javascript method
Instead of doing the previous method before where I write a giant if statement. I wrote it in a javascript if statement saying if the state is true, the persons variable has changed. Then in the return function I call the person variable to print out the whole html div. This not only refactors but its shorter.